feat : PNG generating logic done!

+++ src/utils/utils.cpp
@@ -31,7 +31,7 @@ uint32_t readBigEndian32(std::ifstream& file)
     return (uint32_t(bytes[0]) << 24) | (uint32_t(bytes[1]) << 16) | (uint32_t(bytes[2]) << 8) | uint32_t(bytes[3]);
 }
 
-void write_big_endian_32(std::ofstream& file, uint16_t value)
+void write_big_endian_32(std::ofstream& file, uint32_t value)
 {
     uint8_t bytes[4];
     bytes[0] = (value >> 24) & 0xFF; // Least significant byte LAST (BIG ENDIAN)
